addressing review comments

+++ runtime/server/pbutil/pbutil.go
@@ -54,7 +54,7 @@ func ToValue(v any) (*structpb.Value, error) {
 		if math.IsNaN(v) || math.IsInf(v, 0) {
 			return structpb.NewNullValue(), nil
 		}
-		return structpb.NewNumberValue(float64(v)), nil
+		return structpb.NewNumberValue(v), nil
 	case *big.Int:
 		// Evil cast to float until frontend can deal with bigs:
 		v2, _ := new(big.Float).SetInt(v).Float64()
